logmsg: move tags.{c,h} under logmsg
