+++ lib/Makefile.am
@@ -141,7 +141,6 @@ pkginclude_HEADERS			+= \
 	lib/syslog-names.h		\
 	lib/syslog-ng.h			\
 	lib/string-list.h		\
-	lib/tags.h			\
 	lib/timeutils.h			\
 	lib/tls-support.h		\
 	lib/thread-utils.h		\
@@ -219,7 +218,6 @@ lib_libsyslog_ng_la_SOURCES		= \
 	lib/str-utils.c			\
 	lib/syslog-names.c		\
 	lib/string-list.c		\
-	lib/tags.c			\
 	lib/timeutils.c			\
 	lib/type-hinting.c		\
 	lib/ringbuffer.c		\
