+++ lib/logmsg/Makefile.am
@@ -8,7 +8,8 @@ logmsginclude_HEADERS =     \
  lib/logmsg/nvtable-serialize.h             \
  lib/logmsg/nvtable-serialize-endianutils.h \
  lib/logmsg/tags-serialize.h                \
- lib/logmsg/timestamp-serialize.h
+ lib/logmsg/timestamp-serialize.h           \
+ lib/logmsg/tags.h
 
 logmsg_sources =             \
  lib/logmsg/gsockaddr-serialize.c \
@@ -17,7 +18,8 @@ logmsg_sources =             \
  lib/logmsg/nvtable.c             \
  lib/logmsg/nvtable-serialize.c   \
  lib/logmsg/tags-serialize.c      \
- lib/logmsg/timestamp-serialize.c
+ lib/logmsg/timestamp-serialize.c \
+ lib/logmsg/tags.c
 
 
 include lib/logmsg/tests/Makefile.am
