+++ lib/logmsg/tests/Makefile.am
@@ -3,7 +3,9 @@ lib_logmsg_tests_TESTS =                       \
  lib/logmsg/tests/test_gsockaddr_serialize  \
  lib/logmsg/tests/test_log_message          \
  lib/logmsg/tests/test_logmsg_serialize     \
- lib/logmsg/tests/test_timestamp_serialize
+ lib/logmsg/tests/test_timestamp_serialize  \
+ lib/logmsg/tests/test_tags
+
 
 check_PROGRAMS       += ${lib_logmsg_tests_TESTS}
 
@@ -22,3 +24,6 @@ lib_logmsg_tests_test_logmsg_serialize_LDADD  = $(TEST_LDADD) $(PREOPEN_SYSLOGFO
 
 lib_logmsg_tests_test_nvtable_CFLAGS          = $(TEST_CFLAGS)
 lib_logmsg_tests_test_nvtable_LDADD           = $(TEST_LDADD)
+
+lib_logmsg_tests_test_tags_CFLAGS	      = $(TEST_CFLAGS)
+lib_logmsg_tests_test_tags_LDADD	      = $(TEST_LDADD)
