+++ modules/dbparser/tests/test_patterndb.c
@@ -22,7 +22,6 @@
  */
 
 #include "apphook.h"
-#include "tags.h"
 #include "logmsg/logmsg.h"
 #include "messages.h"
 #include "filter/filter-expr.h"
